[FEAT] Custom components
* color => changed, added
* components.swift -> CustomView
