+++ Money-Planner/Money-Planner/CustomView/Label/DescriptionView.swift
@@ -1,0 +1,26 @@
+//
+//  DescriptionView.swift
+//  Money-Planner
+//
+//  Created by 유철민 on 1/19/24.
+//
+
+import Foundation
+import UIKit
+
+//DescriptionView
+class DescriptionView: MPLabel {
+    // 초기화 메소드
+    init(text: String, alignToCenter: Bool) {
+        super.init(frame: .zero)
+        self.text = text
+        self.textAlignment = alignToCenter ? .center : .left
+        self.font = UIFont.mpFont26B()
+        self.numberOfLines = 0  // 여러 줄 표시 가능
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
